+++ tests/cli_nwr.rs
@@ -34,7 +34,7 @@ fn command_txdb() -> Result<(), Box<dyn std::error::Error>> {
 fn command_ardb() -> Result<(), Box<dyn std::error::Error>> {
     let mut cmd = Command::cargo_bin("nwr")?;
     let output = cmd
-        .arg("txdb")
+        .arg("ardb")
         .arg("--dir")
         .arg("tests/nwr/")
         .output()
@@ -42,7 +42,7 @@ fn command_ardb() -> Result<(), Box<dyn std::error::Error>> {
     let stdout = String::from_utf8(output.stdout).unwrap();
 
     assert!(std::path::Path::new("tests/nwr/ar_refseq.sqlite").exists());
-    assert_eq!(stdout.lines().count(), 8);
+    assert!(stdout.lines().count() > 10);
 
     Ok(())
 }